# api/swagger-json/openapi.json
{"openapi":"3.1.0","info":{"title":"Crawl4AI API","description":"REST API for web crawling and job management","version":"1.0.0"},"paths":{"/api/jobs":{"get":{"tags":["jobs"],"summary":"List Jobs","description":"List all jobs with optional status filter.","operationId":"list_jobs_api_jobs_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by status","title":"Status"},"description":"Filter by status"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Maximum number of jobs to return","default":50,"title":"Limit"},"description":"Maximum number of jobs to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["jobs"],"summary":"Create Job","description":"Create a new crawl job.","operationId":"create_job_api_jobs_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobCreateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/jobs/{job_id}":{"get":{"tags":["jobs"],"summary":"Get Job","description":"Get details of a specific job.","operationId":"get_job_api_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["jobs"],"summary":"Delete Job","description":"Delete a job and its output files.","operationId":"delete_job_api_jobs__job_id__delete","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/jobs/{job_id}/rerun":{"post":{"tags":["jobs"],"summary":"Rerun Job","description":"Re-run an existing job with the same configuration.","operationId":"rerun_job_api_jobs__job_id__rerun_post","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/jobs/{job_id}/results":{"get":{"tags":["jobs"],"summary":"Get Job Results","description":"Get results/files for a completed job.","operationId":"get_job_results_api_jobs__job_id__results_get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobResultsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/schedules":{"get":{"tags":["schedules"],"summary":"List Schedules","description":"List all scheduled jobs.","operationId":"list_schedules_api_schedules_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleListResponse"}}}}}},"post":{"tags":["schedules"],"summary":"Create Schedule","description":"Create a new recurring schedule.","operationId":"create_schedule_api_schedules_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleCreateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/schedules/{schedule_id}":{"get":{"tags":["schedules"],"summary":"Get Schedule","description":"Get details of a specific schedule.","operationId":"get_schedule_api_schedules__schedule_id__get","parameters":[{"name":"schedule_id","in":"path","required":true,"schema":{"type":"integer","title":"Schedule Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["schedules"],"summary":"Delete Schedule","description":"Delete a schedule.","operationId":"delete_schedule_api_schedules__schedule_id__delete","parameters":[{"name":"schedule_id","in":"path","required":true,"schema":{"type":"integer","title":"Schedule Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/schedules/{schedule_id}/toggle":{"put":{"tags":["schedules"],"summary":"Toggle Schedule","description":"Enable or disable a schedule.","operationId":"toggle_schedule_api_schedules__schedule_id__toggle_put","parameters":[{"name":"schedule_id","in":"path","required":true,"schema":{"type":"integer","title":"Schedule Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleToggleRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/schedules/{schedule_id}/run-now":{"post":{"tags":["schedules"],"summary":"Run Schedule Now","description":"Trigger an immediate run of a scheduled job.","operationId":"run_schedule_now_api_schedules__schedule_id__run_now_post","parameters":[{"name":"schedule_id","in":"path","required":true,"schema":{"type":"integer","title":"Schedule Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/settings":{"get":{"tags":["settings"],"summary":"Get Settings","description":"Get all settings.","operationId":"get_settings_api_settings_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SettingsResponse"}}}}}},"put":{"tags":["settings"],"summary":"Update Settings","description":"Update settings.","operationId":"update_settings_api_settings_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SettingsUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SettingsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/settings/test-email":{"post":{"tags":["settings"],"summary":"Send Test Email","description":"Send a test email using configured SMTP settings.","operationId":"send_test_email_api_settings_test_email_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestEmailRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/stats":{"get":{"tags":["stats"],"summary":"Get Stats","description":"Get dashboard statistics.","operationId":"get_stats_api_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatsResponse"}}}}}}},"/":{"get":{"summary":"Root","description":"Root endpoint with API info.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","description":"Health check endpoint.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobCreateRequest":{"properties":{"name":{"type":"string","title":"Name","description":"Job name"},"url":{"type":"string","title":"Url","description":"URL to crawl"},"schema_fields":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Schema Fields","description":"Extraction schema"},"extraction_instruction":{"type":"string","title":"Extraction Instruction","description":"Custom extraction instruction","default":""},"nav_config":{"additionalProperties":true,"type":"object","title":"Nav Config","description":"Navigation config"},"recipients":{"type":"string","title":"Recipients","description":"Email recipients (comma-separated)","default":""},"email_subject":{"type":"string","title":"Email Subject","description":"Email subject","default":""},"run_async":{"type":"boolean","title":"Run Async","description":"Run job immediately after creation","default":true}},"type":"object","required":["name","url"],"title":"JobCreateRequest","description":"Request model for creating a new job."},"JobListResponse":{"properties":{"jobs":{"items":{"$ref":"#/components/schemas/JobResponse"},"type":"array","title":"Jobs"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["jobs","total"],"title":"JobListResponse","description":"Response model for listing jobs."},"JobResponse":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"url":{"type":"string","title":"Url"},"status":{"type":"string","title":"Status"},"created_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created At"},"started_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Started At"},"finished_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Finished At"},"article_count":{"type":"integer","title":"Article Count","default":0},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"},"output_dir":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output Dir"},"config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Config"}},"type":"object","required":["id","name","url","status"],"title":"JobResponse","description":"Response model for a job."},"JobResultsResponse":{"properties":{"job_id":{"type":"string","title":"Job Id"},"files":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Files"},"results":{"anyOf":[{},{"type":"null"}],"title":"Results"}},"type":"object","required":["job_id","files"],"title":"JobResultsResponse","description":"Response model for job results."},"ScheduleCreateRequest":{"properties":{"job_name":{"type":"string","title":"Job Name","description":"Schedule/job name"},"url":{"type":"string","title":"Url","description":"URL to crawl"},"config":{"additionalProperties":true,"type":"object","title":"Config","description":"Job configuration"},"cron":{"type":"string","title":"Cron","description":"Cron expression"},"recipients":{"type":"string","title":"Recipients","description":"Email recipients","default":""},"enabled":{"type":"boolean","title":"Enabled","description":"Whether schedule is enabled","default":true}},"type":"object","required":["job_name","url","cron"],"title":"ScheduleCreateRequest","description":"Request model for creating a schedule."},"ScheduleListResponse":{"properties":{"schedules":{"items":{"$ref":"#/components/schemas/ScheduleResponse"},"type":"array","title":"Schedules"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["schedules","total"],"title":"ScheduleListResponse","description":"Response model for listing schedules."},"ScheduleResponse":{"properties":{"id":{"type":"integer","title":"Id"},"job_name":{"type":"string","title":"Job Name"},"url":{"type":"string","title":"Url"},"cron":{"type":"string","title":"Cron"},"recipients":{"type":"string","title":"Recipients"},"enabled":{"type":"boolean","title":"Enabled"},"last_run":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Run"},"next_run":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Next Run"},"created_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created At"},"config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Config"}},"type":"object","required":["id","job_name","url","cron","recipients","enabled"],"title":"ScheduleResponse","description":"Response model for a schedule."},"ScheduleToggleRequest":{"properties":{"enabled":{"type":"boolean","title":"Enabled"}},"type":"object","required":["enabled"],"title":"ScheduleToggleRequest","description":"Request model for toggling schedule enabled state."},"SettingsResponse":{"properties":{"groq_api_key":{"type":"string","title":"Groq Api Key","default":""},"llm_provider":{"type":"string","title":"Llm Provider","default":"groq/llama-3.1-8b-instant"},"smtp_host":{"type":"string","title":"Smtp Host","default":""},"smtp_port":{"type":"string","title":"Smtp Port","default":"587"},"smtp_user":{"type":"string","title":"Smtp User","default":""},"smtp_password":{"type":"string","title":"Smtp Password","default":""},"default_max_scrolls":{"type":"integer","title":"Default Max Scrolls","default":10},"default_max_inner_pages":{"type":"integer","title":"Default Max Inner Pages","default":5},"default_content_limit":{"type":"integer","title":"Default Content Limit","default":12000}},"type":"object","title":"SettingsResponse","description":"Response model for settings."},"SettingsUpdateRequest":{"properties":{"groq_api_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Groq Api Key"},"llm_provider":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Llm Provider"},"smtp_host":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Smtp Host"},"smtp_port":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Smtp Port"},"smtp_user":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Smtp User"},"smtp_password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Smtp Password"},"default_max_scrolls":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Default Max Scrolls"},"default_max_inner_pages":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Default Max Inner Pages"},"default_content_limit":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Default Content Limit"}},"type":"object","title":"SettingsUpdateRequest","description":"Request model for updating settings."},"StatsResponse":{"properties":{"total_jobs":{"type":"integer","title":"Total Jobs"},"running_jobs":{"type":"integer","title":"Running Jobs"},"completed_jobs":{"type":"integer","title":"Completed Jobs"},"failed_jobs":{"type":"integer","title":"Failed Jobs"},"total_articles":{"type":"integer","title":"Total Articles"},"active_schedules":{"type":"integer","title":"Active Schedules"}},"type":"object","required":["total_jobs","running_jobs","completed_jobs","failed_jobs","total_articles","active_schedules"],"title":"StatsResponse","description":"Response model for dashboard stats."},"SuccessResponse":{"properties":{"success":{"type":"boolean","title":"Success","default":true},"message":{"type":"string","title":"Message","default":""},"data":{"anyOf":[{},{"type":"null"}],"title":"Data"}},"type":"object","title":"SuccessResponse","description":"Generic success response."},"TestEmailRequest":{"properties":{"to_email":{"type":"string","title":"To Email"}},"type":"object","required":["to_email"],"title":"TestEmailRequest","description":"Request model for sending test email."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}